requirements.txt
PyYAML>=6.0
pathlib
